fix tests, remove xclim redundancies
