+++ README.rst
@@ -37,7 +37,7 @@ This package was created with Cookiecutter_ and the `Ouranosinc/cookiecutter-pyp
         :target: https://github.com/psf/black
         :alt: Python Black
 
-.. |logo| image:: https://raw.githubusercontent.com/Ouranosinc/xsdba/main/docs/logos/xclim-logo-small-light.png
+.. |logo| image:: https://raw.githubusercontent.com/Ouranosinc/xsdba/main/docs/logos/xsdba-logo-small-light.png
         :target: https://github.com/Ouranosinc/xsdba
         :alt: Xsdba
         :class: xsdba-logo-small no-theme
