+++ environment-dev.yml
@@ -6,7 +6,7 @@ dependencies:
   - python >=3.9,<3.13
   - boltons
   - bottleneck
-  - cf_xarray
+  - cf_xarray >=0.9.3
   - cftime
   - dask # why was this not installed? This is only pulled in by xclim, optional for xarray.
   - h5netcdf >=1.3.0
@@ -18,8 +18,8 @@ dependencies:
   - statsmodels
   - xarray >=2023.11.0
   - yamale
-  # - xarray >=2022.05.0.dev0
   # Dev tools and testing
+  - netcdf4
   - pip >=24.2.0
   - bump-my-version >=0.25.1
   - watchdog >=4.0.0
