+++ pyproject.toml
@@ -36,7 +36,7 @@ dynamic = ["description", "version"]
 dependencies = [
   "boltons",
   "bottleneck",
-  "cf_xarray",
+  "cf_xarray>=0.9.3",
   "cftime",
   "dask",
   "h5netcdf>=1.3.0",
