+++ src/xsdba/adjustment.py
@@ -1178,7 +1178,7 @@ class NpdfTransform(Adjust):
     The random matrices are generated following a method laid out by :cite:t:`sdba-mezzadri_how_2007`.
 
     This is only part of the full MBCn algorithm, see :ref:`notebooks/sdba:Statistical Downscaling and Bias-Adjustment`
-    for an example on how to replicate the full method with xclim. This includes a standardization of the simulated data
+    for an example on how to replicate the full method with xsdba. This includes a standardization of the simulated data
     beforehand, an initial univariate adjustment and the reordering of those adjusted series according to the rank
     structure of the output of this algorithm.
