+++ tests/test_processing.py
@@ -215,15 +215,15 @@ def test_to_additive(timeseries):
     pr = timeseries(np.array([0, 1e-5, 1, np.e**10]), units="kg m^-2 s^-1")
     prlog = to_additive_space(pr, lower_bound="0 kg m^-2 s^-1", trans="log")
     np.testing.assert_allclose(prlog, [-np.Inf, -11.512925, 0, 10])
-    assert prlog.attrs["sdba_transform"] == "log"
-    assert prlog.attrs["sdba_transform_units"] == "kg m^-2 s^-1"
+    assert prlog.attrs["xsdba_transform"] == "log"
+    assert prlog.attrs["xsdba_transform_units"] == "kg m^-2 s^-1"
 
     with xr.set_options(keep_attrs=True):
         pr1 = pr + 1
     lower_bound = "1 kg m^-2 s^-1"
     prlog2 = to_additive_space(pr1, trans="log", lower_bound=lower_bound)
     np.testing.assert_allclose(prlog2, [-np.Inf, -11.512925, 0, 10])
-    assert prlog2.attrs["sdba_transform_lower"] == 1.0
+    assert prlog2.attrs["xsdba_transform_lower"] == 1.0
 
     # logit
     hurs = timeseries(np.array([0, 1e-3, 90, 100]), units="%")
@@ -234,8 +234,8 @@ def test_to_additive(timeseries):
     np.testing.assert_allclose(
         hurslogit, [-np.Inf, -11.5129154649, 2.197224577, np.Inf]
     )
-    assert hurslogit.attrs["sdba_transform"] == "logit"
-    assert hurslogit.attrs["sdba_transform_units"] == "%"
+    assert hurslogit.attrs["xsdba_transform"] == "logit"
+    assert hurslogit.attrs["xsdba_transform_units"] == "%"
 
     with xr.set_options(keep_attrs=True):
         hursscl = hurs * 4 + 200
@@ -245,8 +245,8 @@ def test_to_additive(timeseries):
     np.testing.assert_allclose(
         hurslogit2, [-np.Inf, -11.5129154649, 2.197224577, np.Inf]
     )
-    assert hurslogit2.attrs["sdba_transform_lower"] == 200.0
-    assert hurslogit2.attrs["sdba_transform_upper"] == 600.0
+    assert hurslogit2.attrs["xsdba_transform_lower"] == 200.0
+    assert hurslogit2.attrs["xsdba_transform_upper"] == 600.0
 
 
 def test_from_additive(timeseries):
